feat(DESIGNSYS-VideoWindow): structural changes
