+++ src/components/VideoWindow/VideoWindow.test.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
 
+
+
 import { VideoWindowProps } from '@components/VideoWindow/types';
 import { render } from '@testing-library/react';
 
-import { VideoWindowV1 } from './subcomponents';
+import VideoWindow from './VideoWindow';
 
 describe('src/components/VideoWindowV1', () => {
   const defaultProps: VideoWindowProps = {
@@ -16,20 +18,20 @@ describe('src/components/VideoWindowV1', () => {
 
   // Тест базового рендера
   test('should render VideoWindowV1 with given title', () => {
-    const { getByText } = render(<VideoWindowV1 {...defaultProps} />);
+    const { getByText } = render(<VideoWindow {...defaultProps} />);
     expect(getByText('Test Video')).toBeInTheDocument();
   });
 
   // Тест отображения видео
   test('should render video element', () => {
-    const { container } = render(<VideoWindowV1 {...defaultProps} />);
+    const { container } = render(<VideoWindow {...defaultProps} />);
     const videoElement = container.querySelector('video');
     expect(videoElement).toBeInTheDocument();
   });
 
   // Тест автостарта видео
   test('should automatically play video when autoPlay is true', () => {
-    const { container } = render(<VideoWindowV1 {...defaultProps} autoPlay={true} />);
+    const { container } = render(<VideoWindow {...defaultProps} autoPlay={true} />);
     const videoElement = container.querySelector('video');
     expect(videoElement).toHaveAttribute('autoplay');
   });
