+++ src/components/VideoWindow/index.ts
@@ -1,3 +1,3 @@
-export * from './VideoWindow';
+export { default as VideoWindow } from './VideoWindow';
 export * from './VideoWindowExample';
 export * from './types';
