+++ src/components/VideoWindow/styles.module.scss
