--- src/components/VideoWindow/subcomponents/VideoWindowV1/index.tsx
@@ -1,1 +1,0 @@
-export * from './VideoWindowV1';
