--- src/components/VideoWindow/subcomponents/index.ts
@@ -1,1 +1,0 @@
-export * from './VideoWindowV1';
